created modules in index.js, fix svg-sprite in the webpack.config
