+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Landing page adaptable for computer, tablet and phone",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "start": "webpack serve",
     "build-dev": "webpack",
@@ -39,6 +40,7 @@
     "sass": "^1.67.0",
     "sass-loader": "^13.3.2",
     "style-loader": "^3.3.3",
+    "svg-sprite-html-webpack": "^2.4.1",
     "webpack": "^5.88.2",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^4.15.1"
