+++ src/partials/header.html
@@ -3,7 +3,7 @@
         <div class="header-wrapper">
             <a href="#" class="logo">
                 <svg class="logo-icon">
-                    <use xlink:href="images/svg/sprite.svg#logo-icon"></use>
+                    <use xlink:href="images/svg/sprites/sprite.svg#logo-icon"></use>
                   </svg>
                 <h1 class="logo-text">Veggieboost</h1>
             </a>
@@ -18,7 +18,7 @@
                     </li>
                     <li class="header-nav-list__item"><a class="header-nav-list__item-link" href="#">
                         <svg class="logo-basket">
-                            <use xlink:href="images/svg/sprite.svg#logo-basket"></use>
+                            <use xlink:href="images/svg/sprites/sprite.svg#logo-basket"></use>
                           </svg>
                         </a></li>
                 </ul>
@@ -26,12 +26,12 @@
             <div class="header-shopping-cart">
                 <a class="header-basket" href="#">
                     <svg class="logo-basket">
-                        <use xlink:href="images/svg/sprite.svg#cart"></use>
+                        <use xlink:href="images/svg/sprites/sprite.svg#cart"></use>
                       </svg>
                 </a>
                 <a href="#" class="header-menu">
                     <svg class="header-menu__icon">
-                        <use xlink:href="images/svg/sprite.svg#header-menu-icon"></use>
+                        <use xlink:href="images/svg/sprites/sprite.svg#header-menu-icon"></use>
                       </svg>
                 </a>
             </div>
