+++ src/partials/section-shipping.html
@@ -2,7 +2,7 @@
     <div class="container">
         <div class="shipping-wrap">
             <svg class="shipping-wrap-img">
-                <use xlink:href="images/svg/sprite.svg#shipping-image"></use>
+                <use xlink:href="images/svg/sprites/sprite.svg#shipping-image"></use>
               </svg>
             <p class="shipping-wrap-text">The best shipping for you! <span class="shipping-wrap-text--first">Your
                     products</span> always arrive in
